into the right structures
